# docs/swagger/swagger.yaml
basePath: /api/v1
definitions:
...
  services.PipelineToggleRequest:
    properties:
      enabled:
        type: boolean
      pipeline_id:
        type: string
    required:
    - pipeline_id
    type: object
  services.ServiceAction:
    enum:
    - start
    - stop
    - restart
    - reload
    - status
    - enable
    - disable
    type: string
    x-enum-varnames:
    - ActionStart
    - ActionStop
    - ActionRestart
    - ActionReload
    - ActionStatus
    - ActionEnable
    - ActionDisable
  services.ServiceActionRequest:
    properties:
      action:
        allOf:
        - $ref: '#/definitions/services.ServiceAction'
        enum:
        - start
        - stop
        - restart
        - reload
        - status
    required:
    - action
    type: object
  services.UpdateSuricataHomeNetRequest:
    properties:
      networks:
        items:
          type: string
        type: array
    required:
    - networks
    type: object
  services.UpdateSuricataInterfaceRequest:
    properties:
      interface:
        type: string
    required:
    - interface
    type: object
host: localhost:8080
info:
  contact:
    email: support@amplifysec.net
    name: API Support
  description: Network Threat Analysis Backend API for managing network interfaces
    and system services
  license:
    name: MIT
    url: http://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: NTA Backend API
  version: "1.0"
paths:
...
  /services/{name}/action:
    post:
      consumes:
      - application/json
      description: Perform an action (start, stop, restart, reload) on a service
      parameters:
      - description: Service name
        in: path
        name: name
        required: true
        type: string
      - description: Action request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/services.ServiceActionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Perform service action
      tags:
      - services
  /services/{name}/status:
    get:
      consumes:
      - application/json
      description: Get the current status of a system service
      parameters:
      - description: Service name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get service status
      tags:
      - services
  /services/logstash/config/check:
    get:
      consumes:
      - application/json
      description: Test Logstash configuration syntax
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Check Logstash configuration
      tags:
      - logstash
  /services/logstash/logs:
    get:
      consumes:
      - application/json
      description: Get recent Logstash service logs
      parameters:
      - description: 'Number of lines (default: 50)'
        in: query
        name: lines
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get Logstash logs
      tags:
      - logstash
  /services/logstash/pipelines:
    get:
      consumes:
      - application/json
      description: List all Logstash pipelines and their status
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get Logstash pipelines
      tags:
      - logstash
  /services/logstash/pipelines/toggle:
    post:
      consumes:
      - application/json
      description: Enable or disable a Logstash pipeline
      parameters:
      - description: Pipeline toggle request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/services.PipelineToggleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Toggle Logstash pipeline
      tags:
      - logstash
  /services/rport/connection:
    get:
      consumes:
      - application/json
      description: Check RPort connection status to remote server
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Check RPort connection
      tags:
      - rport
  /services/rport/logs:
    get:
      consumes:
      - application/json
      description: Get recent RPort service logs
      parameters:
      - description: 'Number of lines (default: 50)'
        in: query
        name: lines
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get RPort logs
      tags:
      - rport
  /services/rsyslog/config/check:
    get:
      consumes:
      - application/json
      description: Check Rsyslog firewall configuration status
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Check Rsyslog configuration
      tags:
      - rsyslog
  /services/rsyslog/config/toggle:
    post:
      consumes:
      - application/json
      description: Enable or disable Rsyslog firewall log collection
      parameters:
      - description: Enable/Disable request
        in: body
        name: request
        required: true
        schema:
          additionalProperties:
            type: boolean
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Toggle Rsyslog firewall configuration
      tags:
      - rsyslog
  /services/rsyslog/firewall-logs:
    get:
      consumes:
      - application/json
      description: Get recent firewall logs
      parameters:
      - description: 'Number of lines (default: 50)'
        in: query
        name: lines
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get Rsyslog firewall logs
      tags:
      - rsyslog
  /services/rsyslog/logs:
    get:
      consumes:
      - application/json
      description: Get recent Rsyslog service logs
      parameters:
      - description: 'Number of lines (default: 50)'
        in: query
        name: lines
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get Rsyslog service logs
      tags:
      - rsyslog
  /services/suricata/alerts:
    get:
      consumes:
      - application/json
      description: Get recent Suricata alerts
      parameters:
      - description: 'Number of alerts (default: 100)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get Suricata alerts
      tags:
      - suricata
  /services/suricata/config/check:
    get:
      consumes:
      - application/json
      description: Test Suricata configuration syntax
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Check Suricata configuration
      tags:
      - suricata
  /services/suricata/home-net:
    get:
      consumes:
      - application/json
      description: Get the current HOME_NET configuration
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get Suricata HOME_NET configuration
      tags:
      - suricata
    put:
      consumes:
      - application/json
      description: Update the HOME_NET configuration
      parameters:
      - description: HOME_NET update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/services.UpdateSuricataHomeNetRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Update Suricata HOME_NET configuration
      tags:
      - suricata
  /services/suricata/interface:
    get:
      consumes:
      - application/json
      description: Get the current network interface being monitored by Suricata
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get Suricata monitoring interface
      tags:
      - suricata
    put:
      consumes:
      - application/json
      description: Change the network interface monitored by Suricata
      parameters:
      - description: Interface update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/services.UpdateSuricataInterfaceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Update Suricata monitoring interface
      tags:
      - suricata
  /services/suricata/logs:
    get:
      consumes:
      - application/json
      description: Get recent Suricata service logs
      parameters:
      - description: 'Number of lines (default: 50)'
        in: query
        name: lines
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get Suricata logs
      tags:
      - suricata
  /services/suricata/rules/update:
    post:
      consumes:
      - application/json
      description: Update Suricata detection rules
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Update Suricata rules
      tags:
      - suricata
  /services/wazuh/logs:
    get:
      consumes:
      - application/json
      description: Get recent Wazuh agent logs
      parameters:
      - description: 'Number of lines (default: 50)'
        in: query
        name: lines
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get Wazuh agent logs
      tags:
      - wazuh
schemes:
- http
- https
swagger: "2.0"